refactor(events): seperate events module
