+++ frontend/src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import {Routes, RouterModule, PreloadAllModules} from '@angular/router';
 import {NotFoundPage} from './auth/pages/not-found/not-found.page';
 
 const routes: Routes = [
-  { path: 'feed', loadChildren: './profile/events/events.module#EventsModule'},
+  { path: 'feed', loadChildren: './events/events.module#EventsModule'},
   { path: 'profile', loadChildren: './profile/profile.module#ProfileModule'},
   { path: '', loadChildren: './auth/auth.module#AuthModule', pathMatch: 'full'},
   { path: '**', component: NotFoundPage}];
