+++ frontend/src/app/events/components/event-card/event-card.component.html
