+++ frontend/src/app/events/pages/event/event.page.scss
