+++ frontend/src/app/events/pages/feed/feed.page.html
@@ -11,7 +11,7 @@
     <ion-item routerLink="/feed">
       Feed
     </ion-item>
-    <ion-item (click)="goToProfile()">
+    <ion-item routerLink="/profile">
       Profile
     </ion-item>
     <ion-item color="danger" (click)="logout()">
