+++ frontend/src/app/events/pages/feed/feed.page.spec.ts
