+++ frontend/src/app/interfaces/user.interface.ts
@@ -2,10 +2,9 @@ import { Object } from './object.interface';
 import {Datetime} from '@ionic/angular';
 
 export interface User extends Object {
+  name: string;
   email: string;
-  profileImage?: string;
-  userDetails: {
-    name: string;
+  details: {
     birth?: Datetime;
     nationality?: string;
     city?: string;
